Added Redis, sorted views and urls in accounts.
Also fixed a bug, when the user first (?query=[nothing]) visited the find_my_friend (search_friends) page, the message "nothing found" was displayed. In this version, when visiting this page for the first time (?query=[nothing]), the message "nothing found" is no longer displayed.
