+++ MathWhiteboard/urls.py
@@ -11,8 +11,8 @@ Class-based views
     1. Add an import:  from other_app.views import Home
     2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
 Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
+    1. Import the include() function: from django.urls_folder import include, path
+    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls_folder'))
 """
 from django.contrib import admin
 from django.urls import path, include
