+++ accounts/urls.py
@@ -1,19 +1,9 @@
-from django.urls import path
-from .views import index, search_friend, send_friend_request, friendship_requests, friendship_request_accept, \
-    friendship_request_reject, UserRegistrationView, send_code, UserEnterRegistrationCodeView, \
-    UserSetPasswordRegistrationView
+from django.urls import path, include
+from .views.other_views import index
 
 app_name = "accounts"
-
 urlpatterns = [
     path("", index, name="index"), # Home page
-    path("find_my_friend/", search_friend, name="friend_search"),
-    path("send_friend_request/<int:receiver_id>", send_friend_request, name="send_friend_request"),
-    path("friendship_requests/", friendship_requests, name="friendship_requests"),
-    path("friendship_accept/<int:invite_id>/", friendship_request_accept, name="friendship_accept"),
-    path("friendship_reject/<int:invite_id>/", friendship_request_reject, name="friendship_reject"),
-    path("register/", UserRegistrationView.as_view(), name="register"),
-    path("send_code/", send_code, name="send_code"),
-    path("enter_registration_code/", UserEnterRegistrationCodeView.as_view(), name="enter_registration_code"),
-    path("registration_set_password/", UserSetPasswordRegistrationView.as_view(), name="registration_set_password")
+    path("", include("accounts.urls_folder.friends_urls")),
+    path("", include("accounts.urls_folder.registration_urls")),
 ]
