+++ accounts/views/other_views.py
@@ -1,0 +1,8 @@
+from django.shortcuts import render
+from django.http import HttpRequest, HttpResponse
+
+# Create your views here.
+def index(request: HttpRequest) -> HttpResponse:
+    return render(request, "base_with_header_menu.html")
+
+
